+++ native/app/core/BungieDefinitions.ts
@@ -1,4 +1,55 @@
-import { object, optional, record, string, number, array, type Output, boolean, unknown } from "valibot";
+import { object, optional, record, string, number, array, type Output, boolean, unknown, value } from "valibot";
+
+// Bungie definition list
+export type DefinitionKey = "DestinySocketCategoryDefinition" | "DestinyStatGroupDefinition";
+
+// -------------------------------
+// The self hosted minified JSON response definition
+// -------------------------------
+
+const MINI_ITEM_DEFINITION_VERSION = 3;
+
+export const ItemResponseSchema = object({
+  helpers: object(
+    {
+      BucketTypeHash: array(number()),
+      DamageTypeHashes: array(number()),
+      Descriptions: array(string()),
+      DisplaySources: array(string()),
+      ExpirationTooltip: array(string()),
+      ExpiredInActivityMessage: array(string()),
+      IconWaterMark: array(string()),
+      InsertionMaterialRequirementHash: array(number()),
+      ItemTypeDisplayName: array(string()),
+      ItemValue: array(number()),
+      PlugCategoryHash: array(number()),
+      PlugCategoryIdentifier: array(string()),
+      ReusablePlugSetHash: array(number()),
+      SingleInitialItemHash: array(number()),
+      SocketCategories: unknown(),
+      SocketCategoryHash: array(number()),
+      SocketEntries: unknown(),
+      SocketIndexes: unknown(),
+      SocketTypeHash: array(number()),
+      StackUniqueLabel: array(string()),
+      StatGroupHash: array(number()),
+      StatHash: array(string()),
+      TalentGridHash: array(number()),
+      TooltipNotifications: array(string()),
+      TraitIds: array(string()),
+      UiItemDisplayStyle: array(string()),
+      UiPlugLabel: array(string()),
+      Versions: array(number()),
+    },
+    unknown(),
+  ),
+
+  items: record(string(), unknown()),
+  version: number([value(MINI_ITEM_DEFINITION_VERSION)]),
+  id: string(),
+});
+
+export type ItemResponse = Output<typeof ItemResponseSchema>;
 
 // -------------------------------
 // All the minified item definitions
