+++ native/app/inventory/UiRowRenderItem.tsx
@@ -1,4 +1,3 @@
-import { type UISections, UISection } from "@/app/bungie/Common.ts";
 import EngramsUI from "@/app/inventory/sections/EngramsUI.tsx";
 import EquipUI from "@/app/inventory/sections/CharacterEquipmentUI.tsx";
 import SeparatorUI from "@/app/inventory/sections/SeparatorUI.tsx";
@@ -7,6 +6,7 @@ import VaultFlexUI from "@/app/inventory/sections/VaultFlexUI.tsx";
 import LostItemsUI from "@/app/inventory/sections/LostItemsUI.tsx";
 import ArtifactUI from "@/app/inventory/sections/ArtifactUI.tsx";
 import VaultSpacerUI from "@/app/inventory/sections/VaultSpacerUI.tsx";
+import { UISection, type UISections } from "@/app/inventory/logic/Helpers.ts";
 
 export const UiCellRenderItem = ({ item }: { item: UISections }) => {
   switch (item.type) {
