+++ native/app/inventory/cells/ArtifactCell.tsx
@@ -1,4 +1,5 @@
-import { DEFAULT_OVERLAP_COLOR, type DestinyIconData } from "@/app/bungie/Common.ts";
+import type { DestinyIconData } from "@/app/bungie/Types.ts";
+import { DEFAULT_OVERLAP_COLOR } from "@/app/inventory/logic/Constants.ts";
 import { useGGStore } from "@/app/store/GGStore.ts";
 import { ICON_SIZE, INNER_FRAME_SIZE } from "@/app/utilities/UISize.ts";
 import { Image } from "expo-image";
