+++ native/app/inventory/cells/DestinyCell.tsx
@@ -1,4 +1,3 @@
-import { CRAFTED_OVERLAY, DEFAULT_OVERLAP_COLOR, type DestinyIconData } from "@/app/bungie/Common.ts";
 import {
   ICON_SIZE,
   INNER_FRAME_SIZE,
@@ -11,6 +10,8 @@ import { useGGStore } from "@/app/store/GGStore.ts";
 import { Image } from "expo-image";
 import React, { useCallback } from "react";
 import { StyleSheet, Text, View, TouchableOpacity } from "react-native";
+import { CRAFTED_OVERLAY, DEFAULT_OVERLAP_COLOR } from "@/app/inventory/logic/Constants.ts";
+import type { DestinyIconData } from "@/app/bungie/Types.ts";
 
 const common = StyleSheet.create({
   quantity: {
