+++ native/app/inventory/pages/GeneralPage.tsx
@@ -1,5 +1,5 @@
+import { InventoryPageEnums } from "@/app/inventory/logic/Helpers.ts";
 import InventoryPage from "@/app/inventory/pages/InventoryPage.tsx";
-import { InventoryPageEnums } from "@/app/bungie/Common.ts";
 
 export default function GeneralPage() {
   return <InventoryPage inventoryPages={InventoryPageEnums.General} />;
