+++ native/app/inventory/pages/WeaponsPage.tsx
@@ -1,4 +1,4 @@
-import { InventoryPageEnums } from "@/app/bungie/Common.ts";
+import { InventoryPageEnums } from "@/app/inventory/logic/Helpers.ts";
 import InventoryPage from "@/app/inventory/pages/InventoryPage.tsx";
 
 export default function WeaponsPage() {
