+++ native/app/inventory/sections/ArtifactUI.tsx
@@ -1,5 +1,5 @@
+import type { DestinyIconData } from "@/app/bungie/Types.ts";
 import ArtifactCell from "@/app/inventory/cells/ArtifactCell.tsx";
-import type { DestinyIconData } from "@/app/bungie/Common.ts";
 import EmptyCell from "@/app/inventory/cells/EmptyCell.tsx";
 import { DEFAULT_SECTION_4_WIDTH, ICON_MARGIN, ICON_SIZE } from "@/app/utilities/UISize.ts";
 import React from "react";
