+++ native/app/inventory/sections/EngramsUI.tsx
@@ -1,8 +1,8 @@
-import { DestinyIconDataEmpty, type DestinyIconData } from "@/app/bungie/Common.ts";
 import { ENGRAMS_SECTION_WIDTH, ENGRAMS_SECTION_HEIGHT, DEFAULT_MARGIN } from "@/app/utilities/UISize.ts";
 import EngramCell from "@/app/inventory/cells/EngramCell.tsx";
 import React from "react";
 import { StyleSheet, View } from "react-native";
+import { DestinyIconDataEmpty, type DestinyIconData } from "@/app/bungie/Types.ts";
 
 const array10 = Array.from({ length: 10 });
