+++ native/app/inventory/sections/LostItemsUI.tsx
@@ -1,4 +1,3 @@
-import type { DestinyIconData } from "@/app/bungie/Common";
 import { DEFAULT_MARGIN, FOOTER_HEIGHT, ICON_MARGIN, ICON_SIZE } from "@/app/utilities/UISize.ts";
 import DestinyCell from "@/app/inventory/cells/DestinyCell";
 import EmptyCell from "@/app/inventory/cells/EmptyCell";
@@ -6,6 +5,7 @@ import { useGGStore } from "@/app/store/GGStore.ts";
 import React from "react";
 import { StyleSheet, View } from "react-native";
 import EngramCell from "@/app/inventory/cells/EngramCell.tsx";
+import type { DestinyIconData } from "@/app/bungie/Types.ts";
 
 type EngramsProps = {
   data: DestinyIconData[];
