+++ native/app/inventory/sections/VaultFlexUI.tsx
@@ -1,9 +1,9 @@
-import type { DestinyIconData } from "@/app/bungie/Common.ts";
 import { DEFAULT_MARGIN, ICON_MARGIN, ICON_SIZE, INV_MAX_WIDTH } from "@/app/utilities/UISize.ts";
 import DestinyCell from "@/app/inventory/cells/DestinyCell.tsx";
 import EmptyCell from "@/app/inventory/cells/EmptyCell.tsx";
 import React, { useMemo } from "react";
 import { StyleSheet, View } from "react-native";
+import type { DestinyIconData } from "@/app/bungie/Types.ts";
 
 type VaultFlexProps = {
   data: DestinyIconData[];
